// resources/js/admin/items/api/enums/item-api-messages.ts
export enum ItemApiMessages {
  Load = 'Unable to load this Item. Please try again.',
  LoadUsage = 'Unable to load Item usage. Please try again.',
  LoadList = 'Unable to load Items. Please try again.',
  LoadFormOptions = 'Unable to load Item form options. Please try again.',
  CreateFallback = 'Unable to create this Item. Please try again.',
  UpdateFallback = 'Unable to update this Item. Please try again.',
  DeleteFallback = 'Unable to delete this Item. Please try again.',
  ImportFallback = 'Unable to import Items. Please try again.',
  ImportSuccess = 'Items imported successfully.',
}